Move constants to its own file to prevent circular dependencies
