+++ src/cfnlint/rules/mappings/KeyName.py
@@ -18,7 +18,7 @@ import re
 import six
 from cfnlint import CloudFormationLintRule
 from cfnlint import RuleMatch
-from cfnlint.helpers import REGEX_ALPHANUMERIC
+from cfnlint.constants import REGEX_ALPHANUMERIC
 
 
 class KeyName(CloudFormationLintRule):
