+++ src/cfnlint/rules/parameters/CidrAllowedValues.py
@@ -18,7 +18,7 @@ import re
 from cfnlint import CloudFormationLintRule
 from cfnlint import RuleMatch
 
-from cfnlint.helpers import REGEX_CIDR
+from cfnlint.constants import REGEX_CIDR
 
 
 class CidrAllowedValues(CloudFormationLintRule):
