+++ src/cfnlint/rules/resources/LimitName.py
@@ -16,7 +16,7 @@
 """
 from cfnlint import CloudFormationLintRule
 from cfnlint import RuleMatch
-from cfnlint.helpers import LIMITS
+from cfnlint.constants import LIMITS
 
 
 class LimitName(CloudFormationLintRule):
