+++ src/cfnlint/rules/resources/ectwo/Vpc.py
@@ -17,8 +17,7 @@
 import re
 from cfnlint import CloudFormationLintRule
 from cfnlint import RuleMatch
-
-from cfnlint.helpers import REGEX_CIDR
+from cfnlint.constants import REGEX_CIDR
 
 
 class Vpc(CloudFormationLintRule):
