+++ src/cfnlint/rules/templates/LimitSize.py
@@ -17,7 +17,7 @@
 import os
 from cfnlint import CloudFormationLintRule
 from cfnlint import RuleMatch
-from cfnlint.helpers import LIMITS
+from cfnlint.constants import LIMITS
 
 
 class LimitSize(CloudFormationLintRule):
